python-backend/pyproject.toml
[project]
name = "python-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "chromadb>=1.3.0",
    "crawl4ai>=0.7.6",
    "ddgs==9.5.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "google-genai>=1.47.0",
    "groq>=0.33.0",
    "httpx>=0.28.1",
    "huggingface-hub[cli]>=0.36.0",
    "langchain>=1.0.3",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.1",
    "lxml>=5.4.0",
    "markdownify>=1.2.2",
    "numpy<2.1.1",
    "ollama>=0.6.0",
    "pydantic>=2.11.7",
    "pypdf>=6.1.3",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "tavily-python>=0.7.12",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.66.5",
    "uvicorn>=0.38.0",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
